test(cli): split app dev hmr tests
